feat : add screens folder and home screen
