+++ src/Screens/Home/style.ts
@@ -1,0 +1,4 @@
+import styled from "styled-components";
+
+export const Header = styled.div``
+export const Content = styled.div`a`
